Update: Homepage Color theme
Un-fucked the color theme for the homepage. I used the coolors page to
generate the color theme for the homepage, like I tried to use just
random colors for the homepage but I realize that doing so is not the
best option as it will fuck me up for the rest of the development time.

So I went with the safe option to generate the colors theme instead.