Initial content of Entry 1
The content of the homepage is supposed to be from the database,
but I opt out of this and just went on putting the content as plain
text inside the index blade of the homepage.

The content of the homepage will be named entry 1, as I planned(I call
the jumbled thoughts in my head a plan) the
first entry for the BlogPost table to be displayed in the homepage.

The content of Entry 1 won't actually be from the database.
This shit is not a good idea, I should have taken into consideration
the complexity of what I was doing first, before laying the ground
work(setting up the database and migration).

The content of Entry 1 is all about the fuck ups that somehow made the
project possible.