Fucking Around With css: Will destroy Homepage
I want to play around the content of the CSS of the HomePage to see
what I can do with it. I am reviewing the contents of the CSS tutorial
made by SuperSimpleDev.

I learned the basics of CSS from him, but I did
not proceed to learn the advanced concepts, I stopped around somewhere
because of various reasons, now I am picking up where I left.

I plan to experiment with the Style of the Homepage. So things will be
messy.

Saved Screenshots of the experiment. Images will be used later for
the content of Entry 1 after the initial design phase.

Which is more of an experimentation rather than an actual design
process.