Styles: Header & Navbar partially complete
The color seem fine but still lifeless. I don't know how to add life to
the nav bar and the header, it still looks dead.

I removed the global selector for adding minimum padding and margin,
I don't think it is a good idea to apply a generic for-everyone
style. The styling looks janky.

I settled for using the `display:grid` for handling the layout of the
homepage, I tried flex but I am having a hard time controlling the
design so I just went with the grid. More straight forward than
`display:flex`.

Also applied the `display:grid` for the body of the homepage, I was
trying to make the nav bar size more responsive to the content,
so I tried forcing the nav bar to change size depending on free
space available.

Which is why I used grid to asign a space for nav bar to grow to.
Zooming in and out will create a different size for the nav bar and the
header.

Well for all the contents of the body, I did not intend for this
effect but I somehow got this effect when I am applying the styles
for the nav bar and the header.