Styles: Main content-spacing & no images
I fixed the spacing of the main content but there are still no images
in the main content. I will with hold adding images until I can
stabilize the styles of the home page and finalize a page design that
can accomodate the images.

There are prepared images, but the ratio of the image and the size is
not optimal for a web page.

The contents of the homepage is nearly complete, there is text,
though I did not add any images, because it would ruin the design.

I tried to make the text size adapt to the size of the screen, relative
sizing, and I am not sure if this is the right approach for
applying text sizes.