Commit msg fix: Added img & content to homepage
Previous commit message is messed up, I did
something wrong, wherein all the changes and edits
I made disappeared so I had to revert to a commit
wherein all the changes are present.

I tried reverting a commit because of a mistake
in creating a commit message, then the problem
appeared. I solved it by resetting the branch to
the commit(`--hard`).

Content added to provide context on the changes
to the project.